+++ src/protobuf-net.Core/Meta/TypeModel.cs
@@ -1,3 +1,4 @@
+
 ﻿using ProtoBuf.Internal;
 using ProtoBuf.Serializers;
 using System;